// LDC.Atlas.WebApp/Client/app/shared/entities/fxdeal-detail.entity.ts
import { PostingStatus } from "../enums/posting-status.enum";

export class FxDealDetail {
    fxDealId: number;
    dataVersionId: number;
    companyId: string;
    reference: string;
    dealDirectionId: number;
    traderId: number;
    contractDate: number;
    maturityDate?: Date;
    departmentId: number;
    fxTradeTypeId?: number;
    currencyCode: string;
    amount: number;
    settlementCurrencyCode: string;
    spotRate: number;
    spotRateType: string;
    fwPoints?: number;
    nominalAccountId: number;
    settlementNominalAccountId: number;
    counterpartyId: number;
    brokerId: number;
    fxDealStatusId: number;
    bankReference: string;
    memorandum: string;
    traderDisplayName: string;
    departmentCode: string;
    accountNumber: string;
    settlementNominalAccountNumber: string;
    totalApplied: number;
    balance: number;
    createdBy: string;
    createdDateTime?: Date;
    modifiedBy: string;
    modifiedDateTime?: Date;
    isEditMode: boolean;
    fxDealDocumentId: number;
    fxSettlementDocumentId: number;
    fxDealDocument: string;
    fxSettlementDocument: string;
    fxDealReverseDocumentId: number;
    fxSettlementReverseDocumentId: number;
    fxDealReverseDocument: string;
    fxSettlementReverseDocument: string;
    fxSettlementDealDocumentPostingStatusId: PostingStatus;
    fxSettlementSettlementDocumentPostingStatusId: PostingStatus;
    ndfAgreedRate: number;
    ndfAgreedDate?: Date;
    fxReference: string;
    c2CCode: string;
    nominalAlternativeAccount: string;
    departmentAlternativeCode: string;
    settlementNominalAccount: string;
    counterpartyCode: string;
    departmentMappingCode: string;
    nominalAccountNumber: string;
    settlementAccountNumber: string;
}